requirements.txt
Django==5.2.8
drf-spectacular
djangorestframework
djangorestframework-simplejwt
requests
